+++ src/streams/receive.rs
@@ -470,6 +470,7 @@ impl StreamRx {
             time,
             header,
             payload,
+            vp8_patch: None,
             nackable: false,
             last_sender_info: self.sender_info.as_ref().map(|l| l.info),
             timestamp: now,